[YS-409] feat: config 파일만 분리하여 복잡한 인터셉터 로직 분리
